+++ app/utils.py
@@ -1,16 +1,38 @@
 import logging
+import sys
+import os
 
 
-def set_up_logging(level=logging.DEBUG):
-    format = "[%(levelname)s] %(asctime)s %(message)s"
-    logging.basicConfig(format=format, level=level)
+class LogHelper(object):
+    @classmethod
+    def get_debug_env_flag(cls):
+        return bool(os.environ.get('DEBUG', 'false').lower() == 'true')
 
+    @classmethod
+    def get_fast_api_logger(cls, name, log_level=logging.DEBUG):
+        """For more on logging, see the readme"""
+        logger = logging.getLogger(name)
+        logger.setLevel(log_level)
+        consoleHandler = logging.StreamHandler(sys.stdout)
+        consoleHandler.setFormatter(logging.Formatter(fmt='%(asctime)s - %(levelname)s - %(message)s'))
+        # log_formatter = logging.Formatter(
+        #     "%(asctime)s [%(processName)s: %(process)d] [%(threadName)s: %(thread)d] [%(levelname)s] %(name)s: %(message)s")
+        # stream_handler.setFormatter(log_formatter)
+        # logger.addHandler(stream_handler)
+        logger.addHandler(consoleHandler)
+        return logger
 
-# pass in __name__
-def get_logger(name):
-    try:
-        from flask import current_app as app
-    except ImportError:
-        print('import error, using a system logger')
-        return logging.getLogger(name)
-    return app.logger
+    @classmethod
+    def get_debug_logger(cls, name):
+        return cls.get_fast_api_logger(name, log_level=logging.DEBUG)
+
+    @classmethod
+    def get_info_logger(cls, name):
+        return cls.get_fast_api_logger(name, log_level=logging.INFO)
+
+    @classmethod
+    def get_env_logger(cls, name):
+        is_debug = cls.get_debug_env_flag()
+        if is_debug:
+            return cls.get_debug_logger(name)
+        return cls.get_info_logger(name)
