+++ docker-compose.yml
@@ -15,6 +15,7 @@ services:
       retries: 5
       start_period: 80s
     environment:
+      - DEBUG=true
       - WLED_IP_LIST=10.0.1.150,10.0.1.179,10.0.1.153
     ports:
       - "9395:9395"
